+++ qcal/__init__.py
@@ -1,6 +1,6 @@
 """Quantum Calibration (qcal) package."""
 
-__version__ = "4.0.0"
+__version__ = "4.0.1"
 
 from qcal.circuit import *
 from qcal.compilation.compiler import Compiler